feat(launcher): 节点生命周期实现体搬进 launcher/nodes.py(原样移动)
统一启动器第 5 步完成。命令面在上一个提交已建,这次搬实现体。

## 用物理移动,不手写重写

system_manager.py 第 90–618 行(NodeConfig / ConfigManager / NODES /
SystemManager,共 529 行)【逐字节移动】到 launcher/nodes.py。

刻意不重写:那 529 行里有大量真机故障攒出来的细节 —— 端口权威解析委派
core.port_config、能力/连接管理器初始化失败时的非致命降级、按组优先级的启动
次序、健康探针的超时与重试……手抄一遍必然丢掉其中几条,而丢掉哪条要等真机上
出问题才知道。移动则由构造保证一样不丢。

搬迁前后做了对照实验:把改动前的 system_manager.py 单独加载,与新版逐字段比
109 个节点记录(id/name/port/group/auto_start/health_check_path/dependencies/
critical/description),【完全一致】。

## 移动唯一会静默改变语义的地方,已逐处修

Path(__file__).parent。实现体原本住在仓库根,那个表达式就是仓库根;搬进
launcher/ 之后它指向 launcher/ —— 两处后果:

  · ConfigManager.CONFIG_FILE 会指向 launcher/config/unified_config.json,
    找不到 → 静默退化成 _get_default_nodes()
  · SystemManager.nodes_dir 会指向 launcher/nodes,那是本模块自己 →
    节点一个都扫不到

两处都不报错,只是扫出空的。已改为显式的 PROJECT_ROOT,并加了一条查真实文件
系统(不是查字符串)的测试:配置文件必须存在、节点目录必须存在、节点数 > 100。

## 下游

- system_manager.py 只剩 CLI 外壳(676 → 167 行),从新家 re-export 四个名字,
  让老命令在并存期继续可用;步骤 8 与其余三个启动器本体一起删。
- health_monitor.py:50 是搬迁前唯一的生产 importer,直接改指 launcher.nodes,
  不经将被删除的外壳中转。

## 两条测试的前提被这次变更作废,已重写而非删除

1. "命令面不该自己起进程" —— nodes.py 现在就是实现体,起子进程是它的职责。
   改为钉搬迁真正要保证的事:全仓只有一份实现(两边不许都有同名类定义)。
2. "import launcher.nodes 不该拉起 system_manager" —— 模块级那次配置读取现在
   在 nodes.py 自己身上。真正要保证的时序没变,换了观察点:
   `main.py --version` 的输出必须只有一行,即没有触发节点表加载。

## 验证

- tests/test_launcher_nodes.py:22 条
- 四个仓库门全过;定向回归 3915 passed / 0 failed
- flake8:launcher/nodes.py 干净;system_manager.py 反而少了 6 条 F401
  (搬走之后那些 import 不再需要);随搬迁带过来的一条既存 E501 顺手修掉
