+++ health_monitor.py
@@ -47,7 +47,9 @@ from fastapi.responses import JSONResponse, PlainTextResponse
 from nodes.common.cors_config import get_cors_origins
 
 # 导入系统管理器
-from system_manager import SystemManager, NODES, NodeConfig
+# 实现体已从 system_manager.py 搬到 launcher/nodes.py（原样移动，非重写）。
+# system_manager.py 在步骤 8 删除；这里直接指向新家，不经它的 re-export 中转。
+from launcher.nodes import NODES, NodeConfig, SystemManager
 
 # SLO 指标
 from core.slo_metrics import get_slo_metrics
@@ -57,38 +59,35 @@ _hm_start_time = time.time()
 
 app = FastAPI(title="Galaxy Health Monitor", version="1.0.0")
 app.add_middleware(
-    CORSMiddleware,
-    allow_origins=get_cors_origins(),
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"]
+    CORSMiddleware, allow_origins=get_cors_origins(), allow_credentials=True, allow_methods=["*"], allow_headers=["*"]
 )
 
 # =============================================================================
 # 健康监控器
 # =============================================================================
 
+
 class HealthMonitor:
     """健康监控器"""
-    
+
     def __init__(self, manager: SystemManager, check_interval: int = 30):
         self.manager = manager
         self.check_interval = check_interval
         self.health_history: Dict[str, List[Dict]] = {}
         self.alert_count: Dict[str, int] = {}
-        
+
         # ===== 集成：初始化能力和连接管理器 =====
         try:
             from core.capability_manager import get_capability_manager
             from core.connection_manager import get_connection_manager
-            
+
             self.capability_manager = get_capability_manager()
             self.connection_manager = get_connection_manager()
         except Exception as e:
             print(f"⚠️  能力/连接管理器初始化失败: {e}")
             self.capability_manager = None
             self.connection_manager = None
-        
+
     async def check_node(self, config: NodeConfig) -> Dict:
         """检查单个节点"""
         t0 = time.monotonic()
@@ -107,69 +106,68 @@ class HealthMonitor:
             "group": config.group,
             "healthy": is_healthy,
             "latency_ms": round(latency_ms, 2),
-            "timestamp": datetime.now().isoformat()
+            "timestamp": datetime.now().isoformat(),
         }
-        
+
         # 记录历史
         if config.id not in self.health_history:
             self.health_history[config.id] = []
-        
+
         self.health_history[config.id].append(status)
-        
+
         # 只保留最近 100 条记录
         if len(self.health_history[config.id]) > 100:
             self.health_history[config.id] = self.health_history[config.id][-100:]
-        
+
         return status
-    
+
     async def check_all(self) -> List[Dict]:
         """检查所有节点"""
         all_configs = []
         for group in NODES.values():
             all_configs.extend(group)
-        
+
         tasks = [self.check_node(config) for config in all_configs]
         results = await asyncio.gather(*tasks)
-        
+
         return results
-    
+
     async def monitor_loop(self):
         """监控循环"""
         print(f"🔍 健康监控已启动（间隔 {self.check_interval} 秒）")
-        
+
         while True:
             try:
                 results = await self.check_all()
-                
+
                 # 统计
                 healthy_count = sum(1 for r in results if r["healthy"])
                 total_count = len(results)
-                
-                print(f"[{datetime.now().strftime('%H:%M:%S')}] "
-                      f"健康: {healthy_count}/{total_count}")
-                
+
+                print(f"[{datetime.now().strftime('%H:%M:%S')}] " f"健康: {healthy_count}/{total_count}")
+
                 # 检查是否需要告警
                 for result in results:
                     if not result["healthy"]:
                         await self.handle_unhealthy_node(result)
-                
+
             except Exception as e:
                 print(f"❌ 监控错误: {e}")
-            
+
             await asyncio.sleep(self.check_interval)
-    
+
     async def handle_unhealthy_node(self, status: Dict):
         """处理不健康的节点"""
         node_id = status["node_id"]
-        
+
         # 增加告警计数
         if node_id not in self.alert_count:
             self.alert_count[node_id] = 0
-        
+
         self.alert_count[node_id] += 1
-        
+
         print(f"⚠️  节点 {status['name']} 不健康（告警次数: {self.alert_count[node_id]}）")
-        
+
         # 如果连续 3 次不健康，尝试重启
         if self.alert_count[node_id] >= 3:
             print(f"🔄 尝试重启节点 {status['name']}...")
@@ -184,23 +182,14 @@ class HealthMonitor:
             except Exception as e:
                 print(f"❌ 节点 {status['name']} 重启异常: {e}")
             self.alert_count[node_id] = 0
-    
+
     def get_summary(self) -> Dict:
         """获取摘要"""
-        summary = {
-            "total_nodes": 0,
-            "healthy_nodes": 0,
-            "unhealthy_nodes": 0,
-            "groups": {}
-        }
-        
+        summary = {"total_nodes": 0, "healthy_nodes": 0, "unhealthy_nodes": 0, "groups": {}}
+
         for group_name, configs in NODES.items():
-            group_summary = {
-                "total": len(configs),
-                "healthy": 0,
-                "unhealthy": 0
-            }
-            
+            group_summary = {"total": len(configs), "healthy": 0, "unhealthy": 0}
+
             for config in configs:
                 if config.id in self.health_history and self.health_history[config.id]:
                     latest = self.health_history[config.id][-1]
@@ -208,12 +197,12 @@ class HealthMonitor:
                         group_summary["healthy"] += 1
                     else:
                         group_summary["unhealthy"] += 1
-            
+
             summary["groups"][group_name] = group_summary
             summary["total_nodes"] += group_summary["total"]
             summary["healthy_nodes"] += group_summary["healthy"]
             summary["unhealthy_nodes"] += group_summary["unhealthy"]
-        
+
         # ===== 集成：添加能力和连接状态 =====
         if self.capability_manager:
             try:
@@ -221,15 +210,16 @@ class HealthMonitor:
                 summary["capabilities"] = cap_stats
             except Exception as e:
                 summary["capabilities"] = {"error": str(e)}
-        
+
         if self.connection_manager:
             try:
                 conn_stats = self.connection_manager.get_stats()
                 summary["connections"] = conn_stats
             except Exception as e:
                 summary["connections"] = {"error": str(e)}
-        
+
         return summary
+
 
 # =============================================================================
 # 全局实例
@@ -242,6 +232,7 @@ monitor = HealthMonitor(manager)
 # API 端点
 # =============================================================================
 
+
 def _get_authority_boundary_status() -> dict:
     """Return a snapshot of the V6 center authority boundary for diagnostics.
 
@@ -251,6 +242,7 @@ def _get_authority_boundary_status() -> dict:
     """
     try:
         from core.center_authority_boundary import evaluate_center_authority_boundary
+
         report = evaluate_center_authority_boundary()
         return {
             "status": "intact" if report.all_domains_intact else "degraded",
@@ -272,16 +264,14 @@ async def root():
         "summary": summary,
     }
 
+
 @app.get("/api/status")
 async def get_status():
     """获取系统状态"""
     results = await monitor.check_all()
     summary = monitor.get_summary()
-    
-    response = {
-        "summary": summary,
-        "nodes": results
-    }
+
+    response = {"summary": summary, "nodes": results}
 
     # Surface authority boundary integrity (V6) for structural diagnostics.
     # This is read-only and non-blocking — it does not affect the request path.
@@ -293,6 +283,7 @@ async def get_status():
     # Surface federation health summary if available
     try:
         from core.galaxy_federation import get_federation, _federation_enabled
+
         fed = get_federation()
         peers = fed.list_peers()
         response["federation"] = {
@@ -308,21 +299,20 @@ async def get_status():
 
     return response
 
+
 @app.get("/api/history/{node_id}")
 async def get_history(node_id: str):
     """获取节点历史"""
     if node_id not in monitor.health_history:
         return {"error": "Node not found"}
-    
-    return {
-        "node_id": node_id,
-        "history": monitor.health_history[node_id]
-    }
+
+    return {"node_id": node_id, "history": monitor.health_history[node_id]}
 
 
 # =============================================================================
 # SLO 指标端点  (PR-G2)
 # =============================================================================
+
 
 @app.get("/metrics", response_class=PlainTextResponse)
 async def prometheus_metrics():
@@ -358,6 +348,7 @@ async def slo_metrics_json():
         }
     """
     return get_slo_metrics().snapshot()
+
 
 @app.get("/api/v1/authority/boundary")
 async def authority_boundary():
@@ -415,12 +406,14 @@ async def health_ready():
 # 启动服务
 # =============================================================================
 
+
 @app.on_event("startup")
 async def startup_event():
     """启动时开始监控循环，并记录启动耗时"""
     startup_ms = (time.time() - _hm_start_time) * 1000.0
     get_slo_metrics().record_startup(startup_ms)
     asyncio.create_task(monitor.monitor_loop())
+
 
 if __name__ == "__main__":
     import argparse
@@ -431,6 +424,7 @@ if __name__ == "__main__":
     # 这里显式解析,让守护进程的调用意图变得诚实可读,并允许覆盖端口。
     try:
         from core.port_config import get_service_port
+
         _default_port = get_service_port("health_monitor")
     except Exception:
         _default_port = 9100
@@ -439,12 +433,15 @@ if __name__ == "__main__":
         description="Galaxy 独立健康看门狗 / Prometheus 导出器(可选;主应用已内置等价端点)"
     )
     _parser.add_argument(
-        "--watchdog", action="store_true",
+        "--watchdog",
+        action="store_true",
         help="以看门狗模式运行(节点巡检 + 自动重启循环由 startup 事件启动;"
-             "本进程即独立看门狗,该标志用于让 daemon 的调用意图显式化)",
+        "本进程即独立看门狗,该标志用于让 daemon 的调用意图显式化)",
     )
     _parser.add_argument(
-        "--port", type=int, default=_default_port,
+        "--port",
+        type=int,
+        default=_default_port,
         help=f"HTTP 监听端口(默认 {_default_port})",
     )
     _parser.add_argument("--host", default="0.0.0.0", help="HTTP 监听地址(默认 0.0.0.0)")
